requirements.txt
Django>=4.1.3
djangorestframework>=3.12.4
psycopg2-binary==2.9.6
djangorestframework-simplejwt>=5.2.2
django-cors-headers==3.14.0